Провести рефакторинг DAO-объектов, исправить тесты

+++ src/main/java/ru/yandex/practicum/storage/genre/GenreDbStorage.java
@@ -4,6 +4,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
 import ru.yandex.practicum.model.Genre;
+import ru.yandex.practicum.storage.mapper.GenreRowMapper;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -12,6 +13,7 @@ import java.util.Optional;
 
 @Repository
 public class GenreDbStorage implements GenreStorage {
+
     private final JdbcTemplate jdbcTemplate;
 
     public GenreDbStorage(JdbcTemplate jdbcTemplate) {
@@ -25,7 +27,7 @@ public class GenreDbStorage implements GenreStorage {
     }
 
     @Override
-    public Optional<Genre> findById(int id) {
+    public Optional<Genre> findById(Long id) {
         String sql = "SELECT * FROM genres WHERE id = ?";
         return jdbcTemplate.query(sql, new GenreRowMapper(), id).stream().findFirst();
     }
