+++ src/main/java/ru/yandex/practicum/storage/mapper/GenreRowMapper.java
@@ -1,6 +1,7 @@
 package ru.yandex.practicum.storage.mapper;
 
 import org.springframework.jdbc.core.RowMapper;
+import ru.yandex.practicum.model.Genre;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -8,6 +9,6 @@ import java.sql.SQLException;
 public class GenreRowMapper implements RowMapper<Genre> {
     @Override
     public Genre mapRow(ResultSet rs, int rowNum) throws SQLException {
-        return new Genre(rs.getInt("id"), rs.getString("name"));
+        return new Genre(rs.getLong("id"), rs.getString("name"));
     }
 }
