+++ src/main/java/ru/yandex/practicum/storage/mpa/MpaDbStorage.java
@@ -4,6 +4,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
 import ru.yandex.practicum.model.MpaRating;
+import ru.yandex.practicum.storage.mapper.MpaRowMapper;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -12,6 +13,7 @@ import java.util.Optional;
 
 @Repository
 public class MpaDbStorage implements MpaStorage {
+
     private final JdbcTemplate jdbcTemplate;
 
     public MpaDbStorage(JdbcTemplate jdbcTemplate) {
@@ -25,7 +27,7 @@ public class MpaDbStorage implements MpaStorage {
     }
 
     @Override
-    public Optional<MpaRating> findById(int id) {
+    public Optional<MpaRating> findById(Long id) {
         String sql = "SELECT * FROM mpa_ratings WHERE id = ?";
         return jdbcTemplate.query(sql, new MpaRowMapper(), id).stream().findFirst();
     }
